requirements.txt
aiogram==2.5.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
Babel==2.8.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
emoji==0.5.4
geojson==2.5.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.24.3
httplib2==0.18.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
lxml==4.4.1
MechanicalSoup==0.12.0
mysql-connector-python==8.0.18
multidict==4.7.3
numpy==1.17.3
oauth2client==4.1.3
Pillow==6.2.1
pipenv==2018.11.26
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyodbc==4.0.27
pyowm==2.10.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.4
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.4
virtualenv-clone==0.5.3
Wikipedia-API==0.5.3
yarl==1.4.2
covid-data-api==1.5 